third_party/blink/web_tests/virtual/popover-hint-new-behavior-disabled/external/wpt/html/semantics/popovers/popover-light-dismiss-hint-expected.txt

This is a testharness.js-based test.
[FAIL] Auto can be nested inside hint (nothing should throw, auto demoted to hint)
  assert_equals: Error, index 2 (innerhint1) expected true but got false
Harness: the test ran to completion.
